implement traceable
